kubernetes/kind/app/requirements.txt
pyyaml>=6.0
pytest>=9.0
pytest-cov>=7.0
